Refactorizando menuhamburger a inyeccion de dependencias y dependencias de abstracciones
