+++ lib/components/menu-hamburguer/MenuHamburguer.type.ts
@@ -1,3 +1,5 @@
+import { ElementOpenAttr } from "lib/shared/class/ElementOpen.csl";
+import { ITemplate } from "lib/shared/interfaces/Template.interface";
 import { ClassNames } from "lib/shared/types";
 
 export type MenuHamburguerType = ClassNames  &{
@@ -7,4 +9,10 @@ export type MenuHamburguerType = ClassNames  &{
   lineDos: string;
   lineTres: string;
   containeChild: string;
+}
+
+export type MenuHamburguerServiceType = {
+  elementAttr: ElementOpenAttr;
+  template: ITemplate<MenuHamburguerType>;
+  eventEmitter: CustomEvent;
 }
